Move xor unit tests
